Updated to .NET 5  and added NullGridTemplate
