+++ Blazm.Components/Blazm.Components/wwwroot/Scripts/OpenInNewWindow.js
@@ -1,0 +1,5 @@
+﻿export function OpenInNewWindow(url, message) {
+    var newwindow = window.open('', '_blank');
+    newwindow.document.write(message);
+    newwindow.location.href = url;
+}
