+++ Blazm.Components/SampleServerSide/Pages/_Host.cshtml
@@ -34,7 +34,6 @@
     </div>
 
     <script src="_framework/blazor.server.js"></script>
-    <script src="_content/Blazm.Components/scripts/jsinterop.min.js"></script>
     <script src="_content/BlazorPro.BlazorSize/blazorSize.min.js"></script>
 </body>
 </html>
